Fix #2143 and fix #2141 CEUR scraping errors CEUR WS scraper did not scrape authors on http://ceur-ws.org/Vol-1191/ CEUR WS scraper now extract pages and computes number of pages An extract error appeared for http://ceur-ws.org/Vol-3240/. This is corrected.
