+++ scholia/utils.py
@@ -109,3 +109,33 @@ def remove_special_characters_url(url):
     encode_string = decoded_url.encode("ascii", "ignore")
     formatted_string = encode_string.decode("utf-8").replace(" ", "")
     return formatted_string
+
+
+def pages_to_number_of_pages(pages):
+    """Compute number of pages based on pages represented as string.
+
+    Parameters
+    ----------
+    pages : str
+        Pages represented as a string.
+
+    Returns
+    -------
+    number_of_pages : int or None
+        Number of pages returned as an integer. If the conversion is not
+        possible then None is returned.
+
+    Examples
+    --------
+    >>> pages_to_number_of_pages('61-67')
+    7
+
+    """
+    number_of_pages = None
+    page_elements = pages.split('-')
+    if len(page_elements) == 2:
+        try:
+            number_of_pages = int(page_elements[1]) - int(page_elements[0]) + 1
+        except ValueError:
+            pass
+    return number_of_pages
